Move commands.NodeAction to enums
