+++ supriya/commands/NodeInfoResponse.py
@@ -1,3 +1,4 @@
+from supriya.enums import NodeAction
 from supriya.commands.Response import Response
 
 
@@ -33,7 +34,7 @@ class NodeInfoResponse(Response):
         import supriya.commands
 
         Response.__init__(self, osc_message=osc_message)
-        self._action = supriya.commands.NodeAction.from_address(action)
+        self._action = NodeAction.from_address(action)
         self._is_group = bool(is_group)
         self._head_node_id = self._coerce_node_id(head_node_id)
         self._next_node_id = self._coerce_node_id(next_node_id)
